Refactor cells
