+++ src/main/kotlin/diary/ui/tabs/notes/cells/SketchCell.kt
@@ -1,0 +1,57 @@
+package diary.ui.tabs.notes.cells
+
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.detectDragGestures
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.unit.dp
+import java.nio.file.Path
+
+class SketchCell(override var name: CellName = CellName("")) : AbstractCell() {
+
+    override fun save(path: Path) {
+        println("Save sketch $path") // TODO
+    }
+
+    @Composable
+    override operator fun invoke() = cell {
+        var offsetX by remember { mutableStateOf(0f) }
+        var offsetY by remember { mutableStateOf(0f) }
+        var action by mutableStateOf<Offset?>(null)
+        val path = androidx.compose.ui.graphics.Path()
+        Canvas(
+            modifier = Modifier
+                .height(250.dp)
+                .fillMaxWidth()
+                .pointerInput(Unit) {
+                    detectDragGestures(
+                        onDragStart = {
+                            offsetX = it.x
+                            offsetY = it.y
+                            path.moveTo(it.x, it.y)
+                        }
+                    ) { _, dragAmount ->
+                        offsetX += dragAmount.x
+                        offsetY += dragAmount.y
+                        path.lineTo(offsetX, offsetY)
+                        action = Offset(offsetX, offsetY)
+                    }
+                }
+        ) {
+            action?.let {
+                drawPath(
+                    path = path,
+                    color = Color(0xFF37596D),
+                    alpha = 1f,
+                    style = Stroke(3f)
+                )
+            }
+        }
+    }
+}
